fix bad cast

+++ src/tetra-receiver.cpp
@@ -228,7 +228,7 @@ auto main(int argc, char** argv) -> int {
       const auto if_gain = result["if"].as<unsigned int>();
       const auto bb_gain = result["bb"].as<unsigned int>();
       const auto& offsets = result["offsets"].as<std::vector<int>>();
-      const auto udp_start = result["udp-start"].as<unsigned int>();
+      const auto udp_start = result["udp-start"].as<uint16_t>();
 
       std::vector<config::Stream> streams;
       const auto input_spectrum = config::SpectrumSlice(center_frequency, sample_rate);
